flex to each news(index 0 - 5)
