+++ components/Article.module.css
@@ -1,0 +1,34 @@
+.body{
+  background-color: yellowgreen;
+  margin: 10px 20px;
+  width: 70vw;
+  margin: 0 auto;
+  padding: 10px 20px;
+}
+
+.news{
+  height: 20vh;
+  border-top: solid 1px blue;
+  border-bottom: solid 1px blue;
+  padding: 5px;
+  overflow-y: scroll;
+}
+
+.news__second{
+  display: inline-block;
+  overflow-y: scroll;
+  width: 30vw;
+}
+
+.news__first{
+  display: inline-block;
+  background-color: black;
+  height: 40vh;
+  width: 30vw;
+}
+
+.news__parent{
+  background-color: aquamarine;
+  display: flex;
+  flex-direction: row;
+}
