+++ components/NewsTimeLine.module.css
@@ -1,3 +1,10 @@
 .body{
-  background-color: grey;
+  /* background-color: grey; */
+  width: 80vw;
+}
+
+@media (max-width: 720px) {
+  .body{
+    width: 100vw;
+  }
 }
