+++ components/NewsTimeLine.tsx
@@ -1,12 +1,11 @@
 import styles from "./NewsTimeLine.module.css"
 import React from 'react';
-import Trend from "./TheTrend";
+import Article from "./Article";
 
 const NewsTimeLine = () => {
   return (
   <div className={styles.body}>
-    body
-    <Trend />
+    <Article />
   </div>
   )
 }
